requirements.txt
strands-agents[anthropic]
strands-agents[bedrock]
